feat: move feed state sealed classes to commons with direct imports
